Remove conditional if statement
Prevented users who have gravatars linked to their emails showing, as the condition user.avatar is false in those instances (they don't have an avatar saved, as it is pulled in from Gravatar) and hence it was being prevented from showing on the site.

There may be a more elegant way but I think it's not needed, anywhere we use the avatar we use a placeholder if one is not provided (e.g. icon on threads/blogs, square space on profiles.